Add InputHandler,instead manualy check for key press and move up or down Spaceship
This make possible to bind the key W and S to move the Alitsa Spaceship Up
or down, because it's has IMovementVertical interface.

Now it's easy to add Atsin Spaceship without to change the code inside of Spaceship
class and only tree line the code can add Atsin and two lines for set
keys for up and down.